+++ packages/devtools/src/runtime/plugins/view/client.ts
@@ -75,6 +75,7 @@ export async function setupDevToolsClient({
     colorMode,
     getIframe,
     updateClient,
+    frameState: state,
   })
 
   let iframe: HTMLIFrameElement | undefined
@@ -242,7 +243,7 @@ export async function setupDevToolsClient({
   // Shortcut to toggle devtools
   addEventListener('keydown', (e) => {
     if (e.code === 'KeyD' && e.altKey && e.shiftKey)
-      client.close()
+      client.toggle()
   })
 
   const app = createApp({